+++ src/main/java/com/snuquill/paperdx/biz/auth/domain/AuditAuthToken.java
@@ -1,22 +1,15 @@
-package com.snuquill.paperdx.admin.auth.domain;
+package com.snuquill.paperdx.biz.auth.domain;
 
-import java.time.Instant;
 import java.time.LocalDateTime;
 
-import org.hibernate.annotations.ColumnDefault;
-
-import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Size;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 @Getter
 @Entity
