+++ src/main/java/com/snuquill/paperdx/biz/auth/domain/AuditAuthTokenRepository.java
@@ -1,4 +1,4 @@
-package com.snuquill.paperdx.admin.auth.domain;
+package com.snuquill.paperdx.biz.auth.domain;
 
 import java.util.Optional;
