+++ src/main/java/com/snuquill/paperdx/biz/auth/domain/vo/AuthTokenHttpHeaders.java
@@ -1,4 +1,4 @@
-package com.snuquill.paperdx.admin.auth.domain.vo;
+package com.snuquill.paperdx.biz.auth.domain.vo;
 
 import org.springframework.http.HttpHeaders;
