+++ src/main/java/com/snuquill/paperdx/biz/auth/service/AuditAuthTokenService.java
@@ -1,12 +1,12 @@
-package com.snuquill.paperdx.admin.auth.service;
+package com.snuquill.paperdx.biz.auth.service;
 
 import java.util.Optional;
 
 import org.springframework.stereotype.Service;
 
-import com.snuquill.paperdx.admin.auth.domain.AuditAuthToken;
-import com.snuquill.paperdx.admin.auth.domain.AuditAuthTokenRepository;
-import com.snuquill.paperdx.admin.auth.domain.vo.AuthTokenPair;
+import com.snuquill.paperdx.biz.auth.domain.AuditAuthToken;
+import com.snuquill.paperdx.biz.auth.domain.AuditAuthTokenRepository;
+import com.snuquill.paperdx.biz.auth.domain.vo.AuthTokenPair;
 import com.snuquill.paperdx.common.execption.DataNotFoundException;
 
 import lombok.RequiredArgsConstructor;
